move obj property to core/
