+++ src/core/lv_obj.h
@@ -22,7 +22,7 @@ extern "C" {
 #include "../misc/lv_area.h"
 #include "../misc/lv_color.h"
 #include "../misc/lv_assert.h"
-#include "../misc/lv_obj_property.h"
+#include "lv_obj_property.h"
 
 /*********************
  *      DEFINES
