+++ src/core/lv_obj_property.c
