+++ src/core/lv_obj_property.h
@@ -14,7 +14,7 @@ extern "C" {
  *      INCLUDES
  *********************/
 #include "../misc/lv_types.h"
-#include "lv_style.h"
+#include "../misc/lv_style.h"
 
 #if LV_USE_OBJ_PROPERTY
